+++ packages/create-modular-react-app/src/__tests__/index.test.ts
@@ -76,7 +76,7 @@ describe('create-modular-react-app', () => {
       │     ├─ package.json
       │     ├─ public
       │     │  ├─ favicon.ico #6pu3rg
-      │     │  ├─ index.html #1wohq3p
+      │     │  ├─ index.html #1m6toxd
       │     │  ├─ logo192.png #1nez7vk
       │     │  ├─ logo512.png #1hwqvcc
       │     │  ├─ manifest.json #19gah8o
@@ -197,7 +197,7 @@ describe('create-modular-react-app', () => {
       │     ├─ package.json
       │     ├─ public
       │     │  ├─ favicon.ico #6pu3rg
-      │     │  ├─ index.html #1wohq3p
+      │     │  ├─ index.html #1m6toxd
       │     │  ├─ logo192.png #1nez7vk
       │     │  ├─ logo512.png #1hwqvcc
       │     │  ├─ manifest.json #19gah8o