+++ packages/modular-scripts/src/start.ts
@@ -2,11 +2,14 @@ import { paramCase as toParamCase } from 'change-case';
 
 import actionPreflightCheck from './utils/actionPreflightCheck';
 import isModularType from './utils/isModularType';
-import execSync from './utils/execSync';
+import execAsync from './utils/execAsync';
 import getLocation from './utils/getLocation';
 import stageView from './utils/stageView';
 import getModularRoot from './utils/getModularRoot';
 import { setupEnvForDirectory } from './utils/setupEnv';
+import { checkBrowsers } from './utils/checkBrowsers';
+import checkRequiredFiles from './utils/checkRequiredFiles';
+import createPaths from './utils/createPaths';
 
 async function start(target: string): Promise<void> {
   const targetPath = await getLocation(target);
@@ -28,25 +31,50 @@ async function start(target: string): Promise<void> {
     startPath = stageView(target);
   } else {
     startPath = targetPath;
+
+    // in the case we're an app then we need to make sure that users have no incorrectly
+    // setup their app folder.
+    const paths = await createPaths(target);
+    await checkRequiredFiles([paths.appHtml, paths.appIndexJs]);
   }
 
-  const startScript = require.resolve(
-    'modular-scripts/react-scripts/scripts/start.js',
-  );
-  const modularRoot = getModularRoot();
-  const targetName = toParamCase(target);
+  await checkBrowsers(startPath);
 
-  execSync('node', [startScript], {
-    cwd: startPath,
-    log: false,
-    // @ts-ignore
-    env: {
-      MODULAR_ROOT: modularRoot,
-      MODULAR_PACKAGE: target,
-      MODULAR_PACKAGE_NAME: targetName,
-    },
-  });
-  return Promise.resolve();
+  // True if there's no preference set - or the preference is for webpack.
+  const useWebpack =
+    !process.env.USE_MODULAR_WEBPACK ||
+    process.env.USE_MODULAR_WEBPACK === 'true';
+
+  // True if the preferene IS set and the preference is esbuid.
+  const useEsbuild =
+    process.env.USE_MODULAR_ESBUILD &&
+    process.env.USE_MODULAR_ESBUILD === 'true';
+
+  // If you want to use webpack then we'll always use webpack. But if you've indicated
+  // you want esbuild - then we'll switch you to the new fancy world.
+  if (!useWebpack || useEsbuild) {
+    const { default: startEsbuildApp } = await import(
+      './esbuild-scripts/start'
+    );
+    await startEsbuildApp(target);
+  } else {
+    const startScript = require.resolve(
+      'modular-scripts/react-scripts/scripts/start.js',
+    );
+    const modularRoot = getModularRoot();
+    const targetName = toParamCase(target);
+
+    await execAsync('node', [startScript], {
+      cwd: startPath,
+      log: false,
+      // @ts-ignore
+      env: {
+        MODULAR_ROOT: modularRoot,
+        MODULAR_PACKAGE: target,
+        MODULAR_PACKAGE_NAME: targetName,
+      },
+    });
+  }
 }
 
 export default actionPreflightCheck(start);